API + SeedData

+++ SoccerFieldBooking.Domain/ValueObjects/FieldSize.cs
@@ -6,17 +6,17 @@ namespace SoccerFieldBooking.Domain.ValueObjects
 {
     public class FieldSize : ValueObject
     {
-        public decimal Length { get; private set; }
-        public decimal Width { get; private set; }
+        public int Length { get; private set; }
+        public int Width { get; private set; }
 
         [NotMapped]
-        public decimal Perimeter { get; private set; }
+        public int Perimeter { get; private set; }
 
         public FieldSize()
         {
         }
 
-        public FieldSize(decimal length, decimal width)
+        public FieldSize(int length, int width)
         {
             Length = length;
             Width = width;
